move minimal_kwargs to common utils
